# api-docs/swagger.yaml
basePath: /api/v1
definitions:
...
  api.NoteCreateRequest:
    properties:
      content:
        minLength: 1
        type: string
      end_time:
        minimum: 0
        type: number
      end_word_index:
        minimum: 0
        type: integer
      quote:
        minLength: 1
        type: string
      start_time:
        minimum: 0
        type: number
      start_word_index:
        description: Use gte=0 so 0 is valid (first word/time); avoid 'required' which
          fails for zero values
        minimum: 0
        type: integer
    required:
    - content
    - quote
    type: object
  api.NoteUpdateRequest:
    properties:
      content:
        minLength: 1
        type: string
    required:
    - content
    type: object
...
  models.JobStatus:
    enum:
    - uploaded
    - pending
    - processing
    - completed
    - failed
    type: string
    x-enum-varnames:
    - StatusUploaded
    - StatusPending
    - StatusProcessing
    - StatusCompleted
    - StatusFailed
  models.MultiTrackFile:
    properties:
      created_at:
        type: string
      file_name:
        description: Original filename (used as speaker name)
        type: string
      file_path:
        description: Full path to audio file
        type: string
      gain:
        description: Gain value from .aup file
        type: number
      id:
        type: integer
      mute:
        description: Whether track is muted
        type: boolean
      offset:
        description: Offset in seconds from .aup file
        type: number
      pan:
        description: Pan value from .aup file (-1.0 to 1.0)
        type: number
      track_index:
        description: Order of the track
        type: integer
      transcription_job:
        allOf:
        - $ref: '#/definitions/models.TranscriptionJob'
        description: Relationships
      transcription_job_id:
        type: string
      updated_at:
        type: string
    type: object
  models.Note:
    properties:
      content:
        description: The user's note content (markdown/plain)
        type: string
      created_at:
        type: string
      end_time:
        type: number
      end_word_index:
        type: integer
      id:
        type: string
      quote:
        description: The exact quoted text chosen by the user
        type: string
      start_time:
        description: Time bounds for the selection (in seconds)
        type: number
      start_word_index:
        description: Indexed selection into transcript by word positions
        type: integer
      transcription:
        allOf:
        - $ref: '#/definitions/models.TranscriptionJob'
        description: Relationships
      transcription_id:
        type: string
      updated_at:
        type: string
    type: object
...
  models.TranscriptionJob:
    properties:
      audio_path:
        type: string
      aup_file_path:
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      diarization:
        type: boolean
      error_message:
        type: string
      id:
        type: string
      individual_transcripts:
        description: JSON-serialized map[string]*string
        type: string
      is_multi_track:
        type: boolean
      merge_error:
        type: string
      merge_status:
        description: none, pending, processing, completed, failed
        type: string
      merged_audio_path:
        type: string
      multi_track_files:
        description: Relationships
        items:
          $ref: '#/definitions/models.MultiTrackFile'
        type: array
      multi_track_folder:
        type: string
      parameters:
        allOf:
        - $ref: '#/definitions/models.WhisperXParams'
        description: WhisperX parameters
      status:
        $ref: '#/definitions/models.JobStatus'
      summary:
        type: string
      title:
        type: string
      transcript:
        type: string
      updated_at:
        type: string
    type: object
...
  models.WhisperXParams:
    properties:
      align_model:
        description: Alignment settings
        type: string
      api_key:
        description: OpenAI settings
        type: string
      attention_context_left:
        description: NVIDIA Parakeet-specific parameters for long-form audio
        type: integer
      attention_context_right:
        type: integer
      batch_size:
        type: integer
      beam_size:
        type: integer
      best_of:
        type: integer
      callback_url:
        description: Webhook settings
        type: string
      chunk_size:
        type: integer
      compression_ratio_threshold:
        type: number
      compute_type:
        type: string
      condition_on_previous_text:
        type: boolean
      device:
        description: Device and computation
        type: string
      device_index:
        type: integer
      diarize:
        description: Diarization settings
        type: boolean
      diarize_model:
        description: 'Options: ''pyannote'', ''nvidia_sortformer'''
        type: string
      fp16:
        type: boolean
      hf_token:
        description: Token and progress
        type: string
      highlight_words:
        type: boolean
      initial_prompt:
        type: string
      interpolate_method:
        type: string
      is_multi_track_enabled:
        description: Multi-track transcription settings
        type: boolean
      language:
        type: string
      length_penalty:
        type: number
      logprob_threshold:
        type: number
      max_line_count:
        type: integer
      max_line_width:
        description: Output formatting
        type: integer
      max_new_tokens:
        description: Voxtral settings
        type: integer
      max_speakers:
        type: integer
      min_speakers:
        type: integer
      model:
        description: Model parameters
        type: string
      model_cache_only:
        type: boolean
      model_dir:
        type: string
      model_family:
        description: Model family (whisper or nvidia)
        type: string
      no_align:
        type: boolean
      no_speech_threshold:
        type: number
      output_format:
        description: Output settings
        type: string
      patience:
        type: number
      print_progress:
        type: boolean
      return_char_alignments:
        type: boolean
      segment_resolution:
        type: string
      speaker_embeddings:
        type: boolean
      suppress_numerals:
        type: boolean
      suppress_tokens:
        type: string
      task:
        description: Task and language
        type: string
      temperature:
        description: Transcription quality settings
        type: number
      temperature_increment_on_fallback:
        type: number
      threads:
        type: integer
      vad_method:
        description: VAD (Voice Activity Detection) settings
        type: string
      vad_offset:
        type: number
      vad_onset:
        type: number
      verbose:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Audio transcription service using WhisperX
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: aScribe API
  version: "1.0"
paths:
...
  /api/v1/notes/{note_id}:
    delete:
      description: Delete a note by its ID
      parameters:
      - description: Note ID
        in: path
        name: note_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete a note
      tags:
      - notes
    get:
      description: Get a note by its ID
      parameters:
      - description: Note ID
        in: path
        name: note_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Note'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Get a note
      tags:
      - notes
    put:
      consumes:
      - application/json
      description: Update the content of a note
      parameters:
      - description: Note ID
        in: path
        name: note_id
        required: true
        type: string
      - description: Note update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.NoteUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Note'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Update a note
      tags:
      - notes
...
  /api/v1/transcription/{id}/notes:
    get:
      description: Get all notes attached to a transcription, ordered by time and
        creation
      parameters:
      - description: Transcription ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Note'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: List notes for a transcription
      tags:
      - notes
    post:
      consumes:
      - application/json
      description: Create a new note attached to the specified transcription
      parameters:
      - description: Transcription ID
        in: path
        name: id
        required: true
        type: string
      - description: Note create payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.NoteCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Note'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Create a note for a transcription
      tags:
      - notes
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: JWT token with Bearer prefix
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"